feat: Add sign up modal to the landing page
This commit adds a sign up modal to the landing page of the RipeMetrics AI Chatbot. The modal is triggered when the "Sign Up" button is clicked and allows users to enter their username and password to sign up for the service. Once the form is submitted, the modal is closed.